requirements/base.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.1.0
django-cors-middleware==1.5.0
django-environ==0.11.2
djangorestframework==3.15.2
drf-yasg==1.21.7

mccabe==0.7.0
packaging==24.1
pycodestyle==2.12.0
pyflakes==3.2.0
pytz==2024.1
PyYAML==6.0.1
setuptools==70.3.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1

argon2-cffi
